requirements.txt
mcp>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pyyaml>=6.0.1
aiofiles>=23.0.0
click>=8.1.0
sse-starlette>=2.0.0
watchfiles>=0.21.0
rich>=13.0.0
requests>=2.31.0